added work with images

+++ Graphic Editor/src/Drawable.h
@@ -7,8 +7,8 @@ public:
     virtual ~Drawable() = default;
 
     // Отрисовка с учетом масштаба и смещения
-    virtual void draw(SDL_Renderer* renderer, float scale, int offsetX, int offsetY) const = 0;
+    virtual void draw(SDL_Renderer* renderer, float scale, float offsetX, float offsetY) const = 0;
 
     // (Необязательно) Проверка попадания в объект — например, для выбора
-    virtual bool contains(int x, int y, float scale, int offsetX, int offsetY) const { return false; }
+    virtual bool contains(int x, int y, float scale, float offsetX, float offsetY) const { return false; }
 };
